Avoid potential use of dangling temporary objects.
In something like

```
  const auto& transformMatrix = detelement.nominal().worldTransformation();
```
nominal() returns a temporary Alignment object by value, and
worldTransformation() returns a reference within that Alignment object.
So we should make a copy.
(The C++ rules for extending the lifetime of a temporary bound to a reference
don't help here because the object bound to a reference is different
from the object whose lifetime needs to be extended.)